Extract `WrapSegment3d` and `WrapSegmentPad` to a distinct file
* Create a new file `wrap_shape1d.cpp` to house wrapping code for 1-D geometry
  entity.
* Move the C++ wrapper classes `WrapSegment3d` and `WrapSegmentPad` from file
  `wrap_World.cpp` to `wrap_shape1d.cpp`.

Claude code command:
* Extract `WrapSegment3d` and `WrapSegmentPad` from `wrap_World.cpp` to a new
  file `wrap_shape1d.cpp`
