+++ cpp/modmesh/universe/pymod/universe_pymod.cpp
@@ -49,6 +49,7 @@ void initialize_universe(pybind11::module & mod)
     {
         wrap_bernstein(mod);
         wrap_shape0d(mod);
+        wrap_shape1d(mod);
         wrap_World(mod);
         wrap_polygon(mod);
     };
