+++ cpp/modmesh/universe/pymod/wrap_World.cpp
@@ -36,480 +36,6 @@ namespace python
 {
 
 template <typename T>
-class MODMESH_PYTHON_WRAPPER_VISIBILITY WrapSegment3d
-    : public WrapBase<WrapSegment3d<T>, Segment3d<T>>
-{
-
-public:
-
-    using base_type = WrapBase<WrapSegment3d<T>, Segment3d<T>>;
-    using wrapped_type = typename base_type::wrapped_type;
-
-    using point_type = typename wrapped_type::point_type;
-    using value_type = typename wrapped_type::value_type;
-
-    friend typename base_type::root_base_type;
-
-protected:
-
-    WrapSegment3d(pybind11::module & mod, char const * pyname, char const * pydoc);
-
-    WrapSegment3d & wrap_management();
-    WrapSegment3d & wrap_operator();
-    WrapSegment3d & wrap_accessor();
-    WrapSegment3d & wrap_geometry();
-}; /* end class WrapSegment3d */
-
-template <typename T>
-WrapSegment3d<T>::WrapSegment3d(pybind11::module & mod, const char * pyname, const char * pydoc)
-    : base_type(mod, pyname, pydoc)
-{
-    namespace py = pybind11;
-
-    (*this)
-        .wrap_management()
-        .wrap_operator()
-        .wrap_accessor()
-        .wrap_geometry()
-        //
-        ;
-}
-
-template <typename T>
-WrapSegment3d<T> & WrapSegment3d<T>::wrap_management()
-{
-    namespace py = pybind11;
-
-    // Constructors.
-    (*this)
-        .def(py::init<point_type const &, point_type const &>(),
-             py::arg("p0"),
-             py::arg("p1"))
-        //
-        ;
-
-    // Python string representations.
-    (*this)
-        .def(
-            "__repr__",
-            [](wrapped_type const & self)
-            {
-                // Hard-code the Python type names in the static variables before finding a systematic way.
-                static char const * stypename = std::is_same_v<T, double> ? "Segment3dFp64" : "Segment3dFp32";
-                static char const * ptypename = std::is_same_v<T, double> ? "Point3dFp64" : "Point3dFp32";
-                return std::format("{}({}({}), {}({}))",
-                                   stypename,
-                                   ptypename,
-                                   self.p0().value_string(),
-                                   ptypename,
-                                   self.p1().value_string());
-            })
-        .def_alias("__repr__", "__str__")
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-WrapSegment3d<T> & WrapSegment3d<T>::wrap_operator()
-{
-    namespace py = pybind11;
-
-    (*this)
-        .def(py::self == py::self) // NOLINT(misc-redundant-expression)
-        .def(py::self != py::self) // NOLINT(misc-redundant-expression)
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-WrapSegment3d<T> & WrapSegment3d<T>::wrap_accessor()
-{
-    namespace py = pybind11;
-
-    (*this)
-        .def(
-            "__len__",
-            [](wrapped_type const & self)
-            { return self.size(); })
-        .def(
-            "__getitem__",
-            [](wrapped_type const & self, size_t it)
-            { return self.at(it); })
-        .def(
-            "__setitem__",
-            [](wrapped_type & self, size_t it, point_type const & vec)
-            { self.at(it) = vec; })
-        //
-        ;
-
-    // p0 (tail) and p1 (head) accessors.
-#define DECL_WRAP(NAME)                               \
-    .def_property(                                    \
-        #NAME,                                        \
-        [](wrapped_type const & self)                 \
-        { return self.NAME(); },                      \
-        [](wrapped_type & self, point_type const & v) \
-        { self.set_##NAME(v); })
-    // clang-format off
-    (*this)
-        DECL_WRAP(p0)
-        DECL_WRAP(p1)
-        ;
-#undef DECL_WRAP
-    // clang-format on
-
-    // x, y, z accessors.
-#define DECL_WRAP(NAME)                                              \
-    .def_property(                                                   \
-        #NAME,                                                       \
-        [](wrapped_type const & self)                                \
-        { return self.NAME(); },                                     \
-        [](wrapped_type & self, typename wrapped_type::value_type v) \
-        { self.set_##NAME(v); })
-    // clang-format off
-    (*this)
-        DECL_WRAP(x0)
-        DECL_WRAP(y0)
-        DECL_WRAP(z0)
-        DECL_WRAP(x1)
-        DECL_WRAP(y1)
-        DECL_WRAP(z1)
-        ;
-#undef DECL_WRAP
-    // clang-format on
-
-    return *this;
-}
-
-template <typename T>
-WrapSegment3d<T> & WrapSegment3d<T>::wrap_geometry()
-{
-    namespace py = pybind11;
-
-    (*this)
-        .def(
-            "mirror",
-            [](wrapped_type & self, std::string const & axis)
-            {
-                if (axis == "x" || axis == "X")
-                {
-                    self.mirror(Axis::X);
-                }
-                else if (axis == "y" || axis == "Y")
-                {
-                    self.mirror(Axis::Y);
-                }
-                else if (axis == "z" || axis == "Z")
-                {
-                    self.mirror(Axis::Z);
-                }
-                else
-                {
-                    throw std::invalid_argument("Segment3d::mirror: axis must be 'x', 'y', or 'z'");
-                }
-            },
-            py::arg("axis"))
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-class MODMESH_PYTHON_WRAPPER_VISIBILITY WrapSegmentPad
-    : public WrapBase<WrapSegmentPad<T>, SegmentPad<T>, std::shared_ptr<SegmentPad<T>>>
-{
-
-public:
-
-    using base_type = WrapBase<WrapSegmentPad<T>, SegmentPad<T>, std::shared_ptr<SegmentPad<T>>>;
-    using wrapped_type = typename base_type::wrapped_type;
-
-    using value_type = typename base_type::wrapped_type::value_type;
-    using point_type = typename base_type::wrapped_type::point_type;
-    using segment_type = typename base_type::wrapped_type::segment_type;
-    using point_pad_type = typename base_type::wrapped_type::point_pad_type;
-
-    friend typename base_type::root_base_type;
-
-protected:
-
-    WrapSegmentPad(pybind11::module & mod, char const * pyname, char const * pydoc);
-
-    WrapSegmentPad & wrap_management();
-    WrapSegmentPad & wrap_accessor_segment();
-    WrapSegmentPad & wrap_accessor_point();
-    WrapSegmentPad & wrap_geometry();
-}; /* end class WrapSegmentPad */
-
-template <typename T>
-WrapSegmentPad<T>::WrapSegmentPad(pybind11::module & mod, const char * pyname, const char * pydoc)
-    : base_type(mod, pyname, pydoc)
-{
-    namespace py = pybind11;
-
-    (*this)
-        .wrap_management()
-        .wrap_accessor_segment()
-        .wrap_accessor_point()
-        .wrap_geometry()
-        //
-        ;
-}
-
-template <typename T>
-WrapSegmentPad<T> & WrapSegmentPad<T>::wrap_management()
-{
-    namespace py = pybind11;
-
-    // Constructors.
-    (*this)
-        .def(
-            py::init(
-                [](uint8_t ndim)
-                { return wrapped_type::construct(ndim); }),
-            py::arg("ndim"))
-        .def(
-            py::init(
-                [](uint8_t ndim, size_t nelem)
-                { return wrapped_type::construct(ndim, nelem); }),
-            py::arg("ndim"),
-            py::arg("nelem"))
-        .def(
-            py::init(
-                [](
-                    SimpleArray<T> & x0,
-                    SimpleArray<T> & y0,
-                    SimpleArray<T> & x1,
-                    SimpleArray<T> & y1,
-                    bool clone)
-                { return wrapped_type::construct(x0, y0, x1, y1, clone); }),
-            py::arg("x0"),
-            py::arg("y0"),
-            py::arg("x1"),
-            py::arg("y1"),
-            py::arg("clone"))
-        .def(
-            py::init(
-                [](
-                    SimpleArray<T> & x0,
-                    SimpleArray<T> & y0,
-                    SimpleArray<T> & z0,
-                    SimpleArray<T> & x1,
-                    SimpleArray<T> & y1,
-                    SimpleArray<T> & z1,
-                    bool clone)
-                { return wrapped_type::construct(x0, y0, z0, x1, y1, z1, clone); }),
-            py::arg("x0"),
-            py::arg("y0"),
-            py::arg("z0"),
-            py::arg("x1"),
-            py::arg("y1"),
-            py::arg("z1"),
-            py::arg("clone"))
-        //
-        ;
-
-    (*this)
-        .def_property_readonly("ndim", &wrapped_type::ndim)
-        .def_timed("pack_array", &wrapped_type::pack_array)
-        .def_timed("expand", &wrapped_type::expand, py::arg("length"))
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-WrapSegmentPad<T> & WrapSegmentPad<T>::wrap_accessor_segment()
-{
-    namespace py = pybind11;
-
-    // Python dunder accessor.
-    (*this)
-        .def("__len__", &wrapped_type::size)
-        .def("__getitem__",
-             [](wrapped_type const & self, size_t it)
-             {
-                 return self.get_at(it);
-             })
-        //
-        ;
-
-    // Append Segment3d.
-    (*this)
-        .def(
-            "append",
-            [](wrapped_type & self, segment_type const & s)
-            {
-                self.append(s);
-            },
-            py::arg("s"))
-        .def(
-            "append",
-            [](wrapped_type & self, point_type const & p0, point_type const & p1)
-            {
-                self.append(p0, p1);
-            },
-            py::arg("p0"),
-            py::arg("p1"))
-        .def(
-            "append",
-            [](wrapped_type & self, value_type x0, value_type y0, value_type x1, value_type y1)
-            {
-                self.append(x0, y0, x1, y1);
-            },
-            py::arg("x0"),
-            py::arg("y0"),
-            py::arg("x1"),
-            py::arg("y1"))
-        .def(
-            "append",
-            [](wrapped_type & self, value_type x0, value_type y0, value_type z0, value_type x1, value_type y1, value_type z1)
-            {
-                self.append(x0, y0, z0, x1, y1, z1);
-            },
-            py::arg("x0"),
-            py::arg("y0"),
-            py::arg("z0"),
-            py::arg("x1"),
-            py::arg("y1"),
-            py::arg("z1"))
-        //
-        ;
-
-    // Extend Segment3d.
-    (*this)
-        .def(
-            "extend_with",
-            [](wrapped_type & self, wrapped_type const & other)
-            {
-                self.extend_with(other);
-            },
-            py::arg("segments"))
-        //
-        ;
-
-    // Additional getter and setter for Segment3d.
-    (*this)
-        .def("get_at",
-             [](wrapped_type const & self, size_t it)
-             {
-                 return self.get_at(it);
-             })
-        .def("set_at",
-             [](wrapped_type & self, size_t it, segment_type const & s)
-             {
-                 self.set_at(it, s);
-             })
-        .def("set_at",
-             [](wrapped_type & self, size_t it, point_type const & p0, point_type const & p1)
-             {
-                 self.set_at(it, p0, p1);
-             })
-        .def("set_at",
-             [](wrapped_type & self, size_t it, value_type x0, value_type y0, value_type x1, value_type y1)
-             {
-                 self.set_at(it, x0, y0, x1, y1);
-             })
-        .def("set_at",
-             [](wrapped_type & self, size_t it, value_type x0, value_type y0, value_type z0, value_type x1, value_type y1, value_type z1)
-             {
-                 self.set_at(it, x0, y0, z0, x1, y1, z1);
-             })
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
-WrapSegmentPad<T> & WrapSegmentPad<T>::wrap_accessor_point()
-{
-    namespace py = pybind11;
-
-    // x, y, z, point element accessors.
-#define DECL_WRAP(NAME)                         \
-    .def(                                       \
-        #NAME,                                  \
-        [](wrapped_type const & self, size_t i) \
-        { return self.NAME(i); })
-    // clang-format off
-    (*this)
-        DECL_WRAP(x0_at)
-        DECL_WRAP(y0_at)
-        DECL_WRAP(z0_at)
-        DECL_WRAP(x1_at)
-        DECL_WRAP(y1_at)
-        DECL_WRAP(z1_at)
-        DECL_WRAP(p0_at)
-        DECL_WRAP(p1_at)
-        ;
-#undef DECL_WRAP
-    // clang-format on
-
-    // x, y, z, point array/batch accessors.
-#define DECL_WRAP(NAME)         \
-    .def_property_readonly(     \
-        #NAME,                  \
-        [](wrapped_type & self) \
-        { return self.NAME(); })
-    // clang-format off
-    (*this)
-        DECL_WRAP(x0)
-        DECL_WRAP(y0)
-        DECL_WRAP(z0)
-        DECL_WRAP(x1)
-        DECL_WRAP(y1)
-        DECL_WRAP(z1)
-        DECL_WRAP(p0)
-        DECL_WRAP(p1)
-        ;
-#undef DECL_WRAP
-    // clang-format on
-
-    return *this;
-}
-
-template <typename T>
-WrapSegmentPad<T> & WrapSegmentPad<T>::wrap_geometry()
-{
-    namespace py = pybind11;
-
-    (*this)
-        .def(
-            "mirror",
-            [](wrapped_type & self, std::string const & axis)
-            {
-                if (axis == "x" || axis == "X")
-                {
-                    self.mirror(Axis::X);
-                }
-                else if (axis == "y" || axis == "Y")
-                {
-                    self.mirror(Axis::Y);
-                }
-                else if (axis == "z" || axis == "Z")
-                {
-                    self.mirror(Axis::Z);
-                }
-                else
-                {
-                    throw std::invalid_argument("SegmentPad::mirror: axis must be 'x', 'y', or 'z'");
-                }
-            },
-            py::arg("axis"))
-        //
-        ;
-
-    return *this;
-}
-
-template <typename T>
 class MODMESH_PYTHON_WRAPPER_VISIBILITY WrapTriangle3d
     : public WrapBase<WrapTriangle3d<T>, Triangle3d<T>>
 {
@@ -1636,10 +1162,6 @@ void wrap_World(pybind11::module & mod)
 {
     WrapBoundBox3d<float>::commit(mod, "BoundBox3dFp32", "BoundBox3dFp32");
     WrapBoundBox3d<double>::commit(mod, "BoundBox3dFp64", "BoundBox3dFp64");
-    WrapSegment3d<float>::commit(mod, "Segment3dFp32", "Segment3dFp32");
-    WrapSegment3d<double>::commit(mod, "Segment3dFp64", "Segment3dFp64");
-    WrapSegmentPad<float>::commit(mod, "SegmentPadFp32", "SegmentPadFp32");
-    WrapSegmentPad<double>::commit(mod, "SegmentPadFp64", "SegmentPadFp64");
     WrapTriangle3d<float>::commit(mod, "Triangle3dFp32", "Triangle3dFp32");
     WrapTriangle3d<double>::commit(mod, "Triangle3dFp64", "Triangle3dFp64");
     WrapTrianglePad<float>::commit(mod, "TrianglePadFp32", "TrianglePadFp32");
